+++ lib/Mutators/AndOrReplacementMutator.cpp
@@ -27,7 +27,7 @@ MutationPoint *AndOrReplacementMutator::getMutationPoint(
     MutationPointAddress &address) {
   if (canBeApplied(*instruction)) {
     std::string diagnostics = "AND-OR Replacement";
-    return new MutationPoint(this, address, instruction, function, diagnostics,
+    return new MutationPoint(this, address, function, diagnostics,
                              sourceLocation, module);
   }
   return nullptr;