+++ lib/Mutators/MathDivMutator.cpp
@@ -25,7 +25,7 @@ MutationPoint *MathDivMutator::getMutationPoint(MullModule *module,
                                                 MutationPointAddress &address) {
   if (canBeApplied(*instruction)) {
     std::string diagnostics = "Math Div: replaced / with *";
-    return new MutationPoint(this, address, instruction, function, diagnostics,
+    return new MutationPoint(this, address, function, diagnostics,
                              sourceLocation, module);
   }