+++ lib/Mutators/ReplaceAssignmentMutator.cpp
@@ -69,8 +69,8 @@ MutationPoint *ReplaceAssignmentMutator::getMutationPoint(
     return nullptr;
   }
 
-  auto mutationPoint = new MutationPoint(this, address, instruction, function,
-                                         diagnostics, sourceLocation, module);
+  auto mutationPoint = new MutationPoint(this, address, function, diagnostics,
+                                         sourceLocation, module);
 
   return mutationPoint;
 }