+++ include/mull/Parallelization/Tasks/MutantExecutionTask.h
@@ -25,7 +25,7 @@ public:
   using Out = std::vector<std::unique_ptr<MutationResult>>;
   using iterator = In::const_iterator;
 
-  MutantExecutionTask(ProcessSandbox &sandbox, Program &program,
+  MutantExecutionTask(const ProcessSandbox &sandbox, Program &program,
                       TestRunner &runner, const Configuration &config,
                       Filter &filter, Mangler &mangler,
                       std::vector<llvm::object::ObjectFile *> &objectFiles,
@@ -37,7 +37,7 @@ public:
 private:
   JITEngine jit;
   Program &program;
-  ProcessSandbox &sandbox;
+  const ProcessSandbox &sandbox;
   TestRunner &runner;
   const Configuration &config;
   Filter &filter;