+++ LLVMCompatibility/7.x.x/LLVMCompatibility.cpp
@@ -52,4 +52,8 @@ void setVersionPrinter(void (*oldPrinter)(),
   llvm::cl::SetVersionPrinter(newPrinter);
 }
 
+Function *GetOrInsertFunction(Module &module, const std::string &name,
+                              FunctionType *functionType) {
+  return cast<Function>(module.getOrInsertFunction(name, functionType));
+}
 } // namespace llvm_compat