+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fixed the original value use-after-free issue #532, #536
  - Automated release process on macOS #537
  - Fix incorrect handling of the NSW/NUW flags #540
+ - Bring back abandoned sandbox approach #541, #542
 
 ## [0.3.0] - 02 Jun 2019